pushed from 1.5

+++ create_cluster.sh
@@ -4,6 +4,7 @@ source "$DIR/common.sh"
 
 config_file="config.json"
 app=
+update_pool=no
 
 function usage(){
     me=`basename $0`
@@ -12,11 +13,12 @@ function usage(){
     echo "$me -c config.json -a APPNAME "
     echo "    -c: config file to use"
     echo "    -a: application (application list)"
+    echo "    -u: [yes, no] update if pool exists, otherwise delete and re-create (default behavior)"
     echo "    -h: help"
     echo ""
 }
 
-while getopts "a: c: h" OPTION
+while getopts "a: c: u: h" OPTION
 do
     case ${OPTION} in
         c)
@@ -24,6 +26,9 @@ do
             ;;
         a)
         app=$OPTARG
+            ;;
+        u)
+        update_pool=$OPTARG
             ;;
         h)
         usage
@@ -36,12 +41,10 @@ shift $(( OPTIND - 1 ));
 
 echo "config file: $config_file"
 echo "app: $app"
+echo "update pool: $update_pool"
 
 read_value subscription_id ".subscription_id"
 read_value location ".location"
-read_value sp_client_id ".service_principal.client_id"
-read_value sp_client_secret ".service_principal.client_secret"
-read_value sp_tenant_id ".service_principal.tenant_id"
 read_value vnet_rg ".infrastructure.network.resource_group"
 read_value vnet ".infrastructure.network.vnet"
 read_value subnet ".infrastructure.network.subnet"
@@ -66,91 +69,65 @@ if [ "$cluster_type" == "batch" ]; then
     read_value batch_rg ".batch.resource_group"
     read_value storage_name ".batch.storage_account"
     read_value container_name ".batch.storage_container"
+    read_value app_storage_rg ".appstorage.resource_group"
+    read_value app_storage_name ".appstorage.storage_account"
+    read_value app_container ".appstorage.app_container"
+    read_value hpc_apps_saskey ".appstorage.app_saskey"
 
-    storage_key=$(az storage account keys list \
-        --resource-group $batch_rg \
-        --account-name $storage_name \
-        --query "[0].value" \
-        --output tsv)
+    # create SAS key
+    create_read_sas_key $batch_rg $storage_name $container_name
+    saskey=$SAS_KEY
+    echo "SAS Key = $saskey"
 
-    # create container for start tasks and policy (if it doesn't exist)
-    container_exists="$(az storage container exists \
-        --account-name $storage_name \
-        --account-key "$storage_key" \
-        --name $container_name --output tsv)"
-    if [ "$container_exists" == "False" ]; then
-        echo "Creating container ($container_name)..."
-        az storage container create \
-            --account-name $storage_name \
-            --account-key "$storage_key" \
-            --name $container_name \
-            --output table
-
-        echo "Adding container policy for read access"
-        az storage container policy create \
-            --container-name $container_name \
-            --name read \
-            --permissions r \
-            --account-key $storage_key \
-            --account-name $storage_name \
-            --start $(date --utc -d "-2 hours" +%Y-%m-%dT%H:%M:%SZ) \
-            --expiry $(date --utc -d "+1 year" +%Y-%m-%dT%H:%M:%SZ)
-    else
-        echo "Container ($container_name) already exists."
+    if [ "$hpc_apps_saskey" == "" ]; then
+        # Create input data SAS KEY
+        create_read_sas_key $app_storage_rg $app_storage_name $app_container
+        hpc_apps_saskey=$SAS_KEY
     fi
+
+    echo "HPC apps SAS Key = $hpc_apps_saskey"
+    hpc_apps_storage_endpoint="https://$app_storage_name.blob.core.windows.net/$app_container"
+    echo "HPC apps storage endpoint is $hpc_apps_storage_endpoint"
 
     # create and upload the node config script (i.e. mount NFS)
     read_value nfsmount ".infrastructure.nfsmount"
     read_value beegfs_mgmt ".infrastructure.beegfs.mgmt"
     read_value beegfs_mount ".infrastructure.beegfs.mount"
 
-    start_task_file_name="starttask-${app}-${timestamp}.sh"
-    cat <<EOF >$start_task_file_name
-#!/bin/bash
+    start_task_file_name="starttask-${app}.sh"
+    temp_start="${timestamp}.sh"
+    cp starttask_common.sh ${temp_start}
 
-mkdir /data
-echo "$nfsserver:${nfsmount}    /data   nfs defaults,proto=tcp,nfsvers=3,rsize=65536,wsize=65536,noatime,soft,timeo=30 0 0" >> /etc/fstab
-chmod 777 /data
-
-mount -a
-
-if [ "$beegfs_mgmt" != "" ]; then
-	sed -i 's/^sysMgmtdHost.*/sysMgmtdHost = '$beegfs_mgmt'/g' /etc/beegfs/beegfs-client.conf
-	echo "$beegfs_mount /etc/beegfs/beegfs-client.conf" > /etc/beegfs/beegfs-mounts.conf
-
-	systemctl enable beegfs-helperd.service
-	systemctl enable beegfs-client.service
-	systemctl start beegfs-helperd.service
-	systemctl start beegfs-client.service
-else
-	systemctl disable beegfs-helperd.service
-	systemctl disable beegfs-client.service
-fi
-EOF
     # append app start task if it exists
-    app_starttask=./apps/$app/starttask_$app.sh
+    app_starttask=$DIR/apps/$app/starttask_$app.sh
     echo $app_starttask
     if [ -f "$app_starttask" ]; then
         echo "Appending to start task ($app_starttask)"
-        cat $app_starttask >> $start_task_file_name
+        cat $app_starttask >> $temp_start
     fi
+
+    # replace keywords in startask
+    replace="s,#NFS_SERVER#,$nfsserver,g"
+    replace+=";s,#NFS_MOUNT#,$nfsmount,g"
+    replace+=";s,#BEEGFS_MGMT#,$beegfs_mgmt,g"
+    replace+=";s,#BEEGFS_MOUNT#,$beegfs_mount,g"
+    replace+=";s,#HPC_APPS_STORAGE_ENDPOINT#,$hpc_apps_storage_endpoint,g"
+    replace+=";s,#HPC_APPS_SASKEY#,$hpc_apps_saskey,g"
+
+    # escape string for sed
+    echo "Replace string: \"$replace\""
+    replace=$(sed "s/\&/\\\&/g" <<< "$replace")
+    sed "$replace" ${temp_start} > ${start_task_file_name}
+
     # upload file
     echo "Uploading start task"
     az storage blob upload \
         --container-name $container_name \
         --account-name $storage_name \
-        --account-key $storage_key \
         --file $start_task_file_name \
         --name $start_task_file_name \
         --output table
-    # create SAS key
-    saskey=$(az storage container generate-sas \
-        --policy-name read \
-        --name $container_name \
-        --account-name $storage_name \
-        --account-key $storage_key \
-        --output tsv)
-    echo "SAS Key = $saskey"
+
     # get storage endpoint
     read_value image_rg ".images.resource_group"
     storage_endpoint=$(az storage account show \
@@ -183,7 +160,8 @@ EOF
         --vnet-name $vnet \
         --name $subnet \
         --query id --output tsv)
-    cat <<EOF >batchpool-${app}-${timestamp}.json
+    batchpool_template=batchpool-${app}.json
+    cat <<EOF >$batchpool_template
 {
   "id": "$pool_id",
   "vmSize": "$vmsize",
@@ -203,11 +181,11 @@ EOF
     "nodeFillType": "Pack"
   },
   "startTask": {
-    "commandLine":"bash starttask-${app}-${timestamp}",
+    "commandLine":"bash ${start_task_file_name}",
     "resourceFiles": [
       {
         "blobSource":"$start_task_source",
-        "filePath":"starttask-${app}-${timestamp}"
+        "filePath":"${start_task_file_name}"
       }
     ],
     "userIdentity": {
@@ -227,39 +205,66 @@ EOF
 
     echo "check if batch pool $pool_id exists"
     batch_pool=$(az batch pool list --filter "id eq '$pool_id'" -o tsv)
-    if [[ ! -z "$batch_pool" ]]; then 
-        echo "deleting batch pool $pool_id"
-        az batch pool delete \
-            --pool-id ${pool_id} \
-            --account-name $batch_account \
-            --yes
+    if [[ ! -z "$batch_pool" ]]; then
+        # if update pool, update it otherwise delete and recreate the pool
+        if [ "$update_pool" = "yes" ]; then 
+            echo "updating batch pool $pool_id"
+            az batch pool set \
+                --pool-id ${pool_id} \
+                --json-file $batchpool_template \
+                --output table
 
-        # wait for the pool to be deleted        
-        loop=true
-        while [ "$loop" != false ]; do
-            sleep 30
-            # retrieve the pool status. If deleting then loop
-            poolstatus=$(az batch pool show --pool-id $pool_id --account-name $batch_account --query "allocationState" --output tsv)
-            echo "pool status is $poolstatus"
-            if [ "$poolstatus" == "deleting" ] || [ "$poolstatus" == "resizing" ] || [ "$poolstatus" == "steady" ]; then
-                continue
+            if [ $? != 0 ]; then
+                echo "ERROR: Failed to update pool $pool_id"
+                exit 1
             fi
-            loop=false
-        done
+        else
+            echo "deleting batch pool $pool_id"
+            az batch pool delete \
+                --pool-id ${pool_id} \
+                --account-name $batch_account \
+                --yes
+
+            # wait for the pool to be deleted        
+            loop=true
+            while [ "$loop" != false ]; do
+                sleep 30
+                # retrieve the pool status. If deleting then loop
+                poolstatus=$(az batch pool show --pool-id $pool_id --account-name $batch_account --query "allocationState" --output tsv)
+                echo "pool status is $poolstatus"
+                if [ "$poolstatus" == "deleting" ] || [ "$poolstatus" == "resizing" ] || [ "$poolstatus" == "steady" ]; then
+                    continue
+                fi
+                loop=false
+            done
+
+            echo "creating batch pool $pool_id"
+            az batch pool create \
+                --json-file $batchpool_template \
+                --output table
+
+            if [ $? != 0 ]; then
+                echo "ERROR: Failed to create pool $pool_id"
+                exit 1
+            fi
+
+        fi
+    else
+        echo "creating batch pool $pool_id"
+        az batch pool create \
+            --json-file $batchpool_template \
+            --output table
+
+        if [ $? != 0 ]; then
+            echo "ERROR: Failed to create pool $pool_id"
+            exit 1
+        fi
     fi
 
-    echo "creating batch pool $pool_id"
-    az batch pool create \
-        --json-file batchpool-${app}-${timestamp}.json \
-        --output table
 
-    if [ $? != 0 ]; then
-        echo "ERROR: Failed to create pool $pool_id"
-        exit 1
-    fi
-
-    rm "batchpool-${app}-${timestamp}.json"
+    rm $batchpool_template
     rm $start_task_file_name
+    rm $temp_start
 
 elif [ "$cluster_type" == "cyclecloud" ]; then
 
@@ -296,38 +301,42 @@ elif [ "$cluster_type" == "cyclecloud" ]; then
         exit 1
     fi
 
-    read_value cyclecloud_admin_password .cyclecloud.admin_password
+    read_secret cyclecloud_admin_password ".cyclecloud.admin_password"
 
     cyclecloud_container_name=$(echo "${cyclecloud_rg}acc01" | tr '[:upper:]' '[:lower:]')
 
     # 2. Check that the CycleCloud server is setup and initialized with the CLI
-    if ! cyclecloud config use $cyclecloud_rg > /dev/null 2>&1 ;then
-        echo "Searching for 'cyclecloud' in Resoruce Group '$cyclecloud_rg'"
+    #if ! cyclecloud config use $cyclecloud_rg > /dev/null 2>&1 ;then
+    if [ "$update_pool" = "yes" ]; then
+        echo "Searching for 'cyclecloud' in Resource Group '$cyclecloud_rg'"
         cyclecloud_fqdn=$(az container show -g $cyclecloud_rg  --name $cyclecloud_container_name  --query "ipAddress.fqdn" | sed 's/\"//g')
         if [ -z $cyclecloud_fqdn ]; then
-            while true; do
-                echo "---- Azure Container Instance $cyclecloud_container_name in resource group $cyclecloud_rg not found."
-                read -p "    Start Azure container running Cyclecloud? [y/n] " yn
-                case $yn in
-                    y )
-                    break;;
-                    n ) exit;;
-                    - ) echo "-- Please answer yes or no.";;
-                esac
-            done
-            echo "Creating resource group $cyclecloud_rg"
-            az group create -n $cyclecloud_rg --location $location 
+
+            group_exists=$(az group exists -n $cyclecloud_rg)
+            if [ "$group_exists" = "false" ];then
+                echo "Creating resource group $cyclecloud_rg"
+                az group create -n $cyclecloud_rg --location $location 
+            fi
+
+            echo "Azure Container Instance $cyclecloud_container_name in resource group $cyclecloud_rg not found."
+            echo "Start Azure container running Cyclecloud"
             echo "Deploying Azure CycleCloud container in resource group $cyclecloud_rg. Please wait. This takes about 5mins."
             hostname_suffix=$(date | md5sum | awk '{print $1}' | cut -c1-8)
             cyclecloud_hostname="cyclecloud$hostname_suffix"
 
-            az container create -g $cyclecloud_rg --location $location --name $cyclecloud_container_name \
-            --dns-name-label $cyclecloud_hostname  --image mcr.microsoft.com/hpc/azure-cyclecloud  \
-            --ip-address public --ports 80 443 --cpu 2 --memory 7 \
-            -e JAVA_HEAP_SIZE=4096 DNS_NAME=$cyclecloud_hostname LOCATION=$location \
-            CYCLECLOUD_AUTOCONFIG=true CYCLECLOUD_ACCEPT_TERMS=true \
-            CYCLECLOUD_USERNAME=$username CYCLECLOUD_PASSWORD=$cyclecloud_admin_password \
-            CYCLECLOUD_USER_PUB_KEY="$rsaPublicKey"
+            az container create -g $cyclecloud_rg --location $location \
+                --name $cyclecloud_container_name \
+                --dns-name-label $cyclecloud_hostname  --image mcr.microsoft.com/hpc/azure-cyclecloud  \
+                --ip-address public --ports 80 443 --cpu 2 --memory 3.5 \
+                -e JAVA_HEAP_SIZE=3072 DNS_NAME=$cyclecloud_hostname LOCATION=$location \
+                CYCLECLOUD_AUTOCONFIG=true CYCLECLOUD_ACCEPT_TERMS=true \
+                CYCLECLOUD_USERNAME=$username CYCLECLOUD_PASSWORD=$cyclecloud_admin_password \
+                CYCLECLOUD_USER_PUB_KEY="$rsaPublicKey"
+
+            if [ $? != 0 ]; then
+                echo "ERROR: Failed to create container $cyclecloud_container_name in $location"
+                exit 1
+            fi
 
             cyclecloud_fqdn=$(az container show -g $cyclecloud_rg  --name $cyclecloud_container_name  --query "ipAddress.fqdn" | sed 's/\"//g')
             echo "Waiting for Azure CycleCloud container at $cyclecloud_fqdn to start "
@@ -339,11 +348,17 @@ elif [ "$cluster_type" == "cyclecloud" ]; then
             sleep 10s
 
         fi
+
         cyclecloud_fqdn=$(az container show -g $cyclecloud_rg  --name $cyclecloud_container_name  --query "ipAddress.fqdn" | sed 's/\"//g')
         echo "CycleCloud server found: $cyclecloud_fqdn"
         echo "Initializing CLI with CycleCloud server"
-        read_value cyclecloud_admin_user .cyclecloud.admin_user
-        cyclecloud initialize --force --batch --name $cyclecloud_rg --url=https://$cyclecloud_fqdn --verify-ssl=false --username=$cyclecloud_admin_user --password=$cyclecloud_admin_password
+        #read_value cyclecloud_admin_user .cyclecloud.admin_user
+        cyclecloud initialize --force --batch --name $cyclecloud_rg --url=https://$cyclecloud_fqdn --verify-ssl=false --username=$username --password=$cyclecloud_admin_password
+
+        read_value sp_client_id ".service_principal.client_id"
+        read_secret sp_client_secret ".service_principal.client_secret"
+        read_value sp_tenant_id ".service_principal.tenant_id"
+        read_value storage_name ".cyclecloud.storage_account"
 
         cat > azure_account.json << FILE
 {
@@ -358,7 +373,7 @@ elif [ "$cluster_type" == "cyclecloud" ]; then
     "Name": "azure",
     "Provider": "azure",
     "ProviderId": "$subscription_id",
-    "RMStorageAccount": "$cyclecloud_hostname",
+    "RMStorageAccount": "$storage_name",
     "RMStorageContainer": "cyclecloud"
 }
 FILE
@@ -442,7 +457,7 @@ EOF
     echo ""
     echo "  Login to the web interface using your username: '$(whoami)' "
     echo "  The password is the secret of your service-principal"
-    echo "  You can find the service principal-sercret in the config file: ./$config_file"
+    echo "  You can find the service principal-secret in the config file: $config_file"
     echo ""
     echo "  To see running clusters and the ipaddress of the cluster nodes, run: "
     echo "      $ $HOME/bin/cyclecloud show_cluster -l to see running clusters"
